File restructuring
- separated parser.hpp -> btstream.{hpp,cpp}
- updated cmake
